LPS-106397 Move OrderByExpression and OrderByInfo into sort package.
