+++ modules/core/petra/petra-sql-dsl/src/main/java/com/liferay/petra/sql/dsl/query/sort/OrderByExpression.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.petra.sql.dsl.query;
+package com.liferay.petra.sql.dsl.query.sort;
 
 import com.liferay.petra.sql.dsl.ast.ASTNode;
 import com.liferay.petra.sql.dsl.expression.Expression;
