+++ modules/core/petra/petra-sql-dsl/src/main/resources/com/liferay/petra/sql/dsl/query/sort/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
