+++ modules/core/petra/petra-sql-dsl/src/test/java/com/liferay/petra/sql/dsl/ast/SQLDSLTest.java
@@ -42,8 +42,6 @@ import com.liferay.petra.sql.dsl.query.GroupByStep;
 import com.liferay.petra.sql.dsl.query.HavingStep;
 import com.liferay.petra.sql.dsl.query.JoinStep;
 import com.liferay.petra.sql.dsl.query.LimitStep;
-import com.liferay.petra.sql.dsl.query.OrderByExpression;
-import com.liferay.petra.sql.dsl.query.OrderByInfo;
 import com.liferay.petra.sql.dsl.query.OrderByStep;
 import com.liferay.petra.sql.dsl.query.impl.From;
 import com.liferay.petra.sql.dsl.query.impl.GroupBy;
@@ -52,13 +50,15 @@ import com.liferay.petra.sql.dsl.query.impl.Join;
 import com.liferay.petra.sql.dsl.query.impl.JoinType;
 import com.liferay.petra.sql.dsl.query.impl.Limit;
 import com.liferay.petra.sql.dsl.query.impl.OrderBy;
-import com.liferay.petra.sql.dsl.query.impl.OrderByExpressionImpl;
 import com.liferay.petra.sql.dsl.query.impl.QueryExpression;
 import com.liferay.petra.sql.dsl.query.impl.QueryTable;
 import com.liferay.petra.sql.dsl.query.impl.Select;
 import com.liferay.petra.sql.dsl.query.impl.SetOperation;
 import com.liferay.petra.sql.dsl.query.impl.SetOperationType;
 import com.liferay.petra.sql.dsl.query.impl.Where;
+import com.liferay.petra.sql.dsl.query.sort.OrderByExpression;
+import com.liferay.petra.sql.dsl.query.sort.OrderByInfo;
+import com.liferay.petra.sql.dsl.query.sort.impl.OrderByExpressionImpl;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.test.rule.CodeCoverageAssertor;
